requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
jinja2